+++ flexmeasures/data/scripts/cli_tasks/testing.py
@@ -22,8 +22,14 @@ from flexmeasures.data.services.forecasting import (
     handle_forecasting_exception,
 )
 
+"""
+These functions are meant for FlexMeasures developers to manually test some internal 
+functionality.
+They are not registered as app command per default, as we don't need to show them to users. 
+"""
 
-@app.cli.command()
+# un-comment to use as CLI function
+# @app.cli.command()
 def test_making_forecasts():
     """
     Manual test to enqueue and process a forecasting job via redis queue
@@ -70,7 +76,8 @@ def test_making_forecasts():
     )
 
 
-@app.cli.command()
+# un-comment to use as CLI function
+# @app.cli.command()
 @click.option("--asset-type", help="Asset type name.")
 @click.option("--asset", help="Asset name.")
 @click.option(